Hint that the user should type into the box
I believe this is the first time the user has to actually create a program on their own. I'm an experienced coder (40 years) and didn't realise I was supposed to type into the 'Question' box, so clicked 'Answer' (I'm supposed to type my answer there, right?). I clicked the 'Run' button... the code executed... and then I realised I was already looking at the solution. Hmm. 

So here's my *minimal* 'hint' for the user to clue them in to where they should type.