Improving variable names
The paragraph just above here tells the user about the importance of choosing good variable names... but the variables in the next exercise are ambiguous. 

Does 'apples' mean the number of apples, or the cost of an apple? I assumed the latter and changed the program to 
apples = 0.4
pears = 0.65
totalCost = apples + pears

AND had my expectations 'confirmed' when I found that option A in the multi-choice below matched my code. Very surprised was I to find I was wrong!

Let's avoid this confusion by taking the timely opportunity to differentiate between the two apple-related variables. This is the perfect time because the spreadsheet above has visually 'clued' the user in to the possibility of a real-world object have multiple attributes that must be tracked as variables. numApples & costPerApple are not interchangeable, so no confusion, and they follow the same format (adjectiveNoun) that we've been following so far (numPeople, totalCost, unitPrice). The user needs to realise that short yet specific variable names are one of the keys to managing complexity and thinking clearly about your code.